+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/servlets/action/CreateLabelServletTest.java
@@ -18,6 +18,11 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Map;
 
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_KEY;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_MESSAGE;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_MESSAGEENTRY;
+import static junit.framework.Assert.assertNull;
+import static org.apache.jackrabbit.JcrConstants.JCR_PRIMARYTYPE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -84,8 +89,33 @@ class CreateLabelServletTest {
         Resource resource = context.resourceResolver().getResource("/content/dictionaries/i18n/en/greeting");
         ValueMap properties = resource.getValueMap();
         assertNotNull(resource);
-        assertEquals("sling:MessageEntry", properties.get("jcr:primaryType"));
-        assertEquals("greeting", properties.get("sling:key"));
-        assertEquals("Hello", properties.get("sling:message"));
+        assertEquals(SLING_MESSAGEENTRY, properties.get(JCR_PRIMARYTYPE));
+        assertEquals("greeting", properties.get(SLING_KEY));
+        assertEquals("Hello", properties.get(SLING_MESSAGE));
+    }
+
+    @Test
+    void doPostWithEmptyMessage() throws ServletException, IOException {
+        context.create().resource("/content/dictionaries/i18n/en", Map.of("jcr:language", "en"));
+        context.create().resource("/content/dictionaries/i18n/fr", Map.of("jcr:language", "fr"));
+
+        context.request().setMethod("POST");
+        context.request().setParameterMap(Map.of(
+                "dictionary", "/content/dictionaries/i18n",
+                "key", "greeting",
+                "en", "Hello",
+                "fr", ""
+        ));
+
+        servlet.service(context.request(), context.response());
+
+        assertEquals(HttpServletResponse.SC_OK, context.response().getStatus());
+
+        Resource resource = context.resourceResolver().getResource("/content/dictionaries/i18n/fr/greeting");
+        ValueMap properties = resource.getValueMap();
+        assertNotNull(resource);
+        assertEquals(SLING_MESSAGEENTRY, properties.get(JCR_PRIMARYTYPE));
+        assertEquals("greeting", properties.get(SLING_KEY));
+        assertNull(properties.get(SLING_MESSAGE));
     }
 }